requirements.txt
PasteDeploy==1.5.2
pyramid==1.5.7
python-postmark-inbound==1.1.0
repoze.lru==0.6
translationstring==1.3
venusian==1.0
WebOb==1.4.1
wheel==0.24.0
zope.deprecation==4.1.2
zope.interface==4.1.2
